requirements.txt
google_api_python_client==2.66.0
google_auth_oauthlib==0.4.6
hausdorff==0.2.6
keras==2.9.0
matplotlib==3.2.2
MedPy==0.4.0
nibabel==3.0.2
numpy==1.21.6
opencv_contrib_python==4.6.0.66
pandas==1.3.5
protobuf==4.21.9
scikit_learn==1.1.3
tensorflow==2.9.2
tqdm==4.64.1
lungmask